+++ blueprint/Brocfile.js
@@ -14,7 +14,10 @@ var app = new EmberApp({
 });
 
 // Use this to add additional libraries to the generated output files.
-app.import('vendor/ember-data/ember-data.js');
+app.import({
+  development: 'vendor/ember-data/ember-data.js',
+  production:  'vendor/ember-data/ember-data.prod.js'
+});
 
 // If the library that you are including contains AMD or ES6 modules that
 // you would like to import into your application please specify an