+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # ember-cli Changelog
 
+* [ENHANCEMENT] Use ember-data.prod.js when ENV=production [#909](https://github.com/stefanpenner/ember-cli/pull/909).
+
 ### 0.0.29
 
 * [ENHANCEMENT] less CPU intensive watching thanks to @krisselden's https://github.com/krisselden/broccoli-sane-watcher and @amasad's https://github.com/amasad/sane